// react-frontend/src/types/nfeTypes.ts
export interface NFeIdentification {
    internalKey: string;
    cUF: string;
    cNF: string;
    natOp: string;
    mod: string;
    serie: string;
    nNF: string;
    dhEmi: string;
    dhSaiEnt?: string;
    tpNF: string;
    idDest: string;
    cMunFG: string;
    tpImp: string;
    tpEmis: string;
    cDV: string;
    tpAmb: string;
    finNFe: string;
    indFinal: string;
    indPres: string;
    indIntermed?: string;
    procEmi: string;
    verProc: string;
    dhCont?: string;
    xJust?: string;
}